+++ src/sinks/azure_blob/request_builder.rs
